Split gui.py into focused package modules
The single mkpfs/gui.py file had grown to 1843 lines mixing translations,
theme, widgets, and six distinct operation panels. This refactor splits it
into a gui/ package with clear separation of concerns:

- gui/i18n.py — translations and locale state
- gui/theme.py — palette, fonts, and panel accent color
- gui/widgets.py — reusable UI components
- gui/panels/ — one module per operation (pack_folder, pack_file,
  verify, inspect, tree, unpack) plus a shared base class
- gui/app.py — MkPFSApp and main() entry point
- gui/__init__.py — re-exports main() for the mkpfs-gui entry point

No user-facing changes. The old gui.py is deleted.
