+++ mkpfs/gui/__init__.py
@@ -1,0 +1,3 @@
+"""Graphical user interface for mkpfs - futuristic neon glassmorphism design."""
+
+from .app import main
