+++ mkpfs/gui/i18n.py
@@ -1,0 +1,331 @@
+"""Internationalisation state and translation lookup for the mkpfs GUI."""
+
+# ---------------------------------------------------------------------------
+# Translation tables
+# ---------------------------------------------------------------------------
+
+_TRANSLATIONS: dict[str, dict[str, str]] = {
+    "en": {
+        # App chrome
+        "app_subtitle": "PlayStation File System",
+        "lang_label": "Language",
+        "operations": "OPERATIONS",
+        "version_footer": "v{version}",
+        # Nav labels
+        "nav_pack_folder": "📦  Pack Folder",
+        "nav_pack_file": "📄  Pack File",
+        "nav_verify": "✅  Verify",
+        "nav_inspect": "🔍  Inspect",
+        "nav_tree": "🌲  Tree",
+        "nav_unpack": "📂  Unpack",
+        # Common
+        "paths": "Paths",
+        "options": "Options",
+        "encryption": "Encryption",
+        "output_log": "Output Log",
+        "export_log": "Export Log",
+        "run": "Run",
+        "running": "Running…",
+        "browse": "Browse",
+        "ok": "✓ Completed successfully.",
+        "err_process": "✗ Process exited with code {}.",
+        "err_not_found": "✗ mkpfs not found. Is the virtual environment active?",
+        "err_unexpected": "[ERROR] Unexpected: {}",
+        # Pack Folder
+        "pf_title": "Pack Folder",
+        "pf_subtitle": "Build a PFS image from a source directory",
+        "pf_src_label": "Source Folder",
+        "pf_src_ph": "Select source folder…",
+        "pf_out_label": "Output Image",
+        "pf_out_ph": "Select output path…",
+        "pf_version": "Version",
+        "pf_compress": "PFSC Compression",
+        "pf_signed": "Signed Image",
+        "pf_verify": "Verify After Pack",
+        "pf_dry": "Dry Run",
+        "pf_temp": "Temp Folder (optional)",
+        "pf_temp_ph": "Custom temp folder…",
+        "pkf_temp": "Temp Folder (optional)",
+        "pkf_temp_ph": "Custom temp folder…",
+        "pf_err_src": "✗ Source folder is required.",
+        "pf_err_out": "✗ Output image path is required.",
+        # Pack File
+        "pkf_title": "Pack File",
+        "pkf_subtitle": "Wrap a single file into a PFSC compressed image",
+        "pkf_src_label": "Source File",
+        "pkf_src_ph": "Select source file…",
+        "pkf_out_label": "Output Image",
+        "pkf_out_ph": "Select output path…",
+        "pkf_version": "Version",
+        "pkf_compress": "PFSC Compression",
+        "pkf_err": "✗ Source file and output path are required.",
+        # Verify
+        "v_title": "Verify",
+        "v_subtitle": "Validate image structure, inodes, and payload checksums",
+        "v_image_label": "PFS Image",
+        "v_image_ph": "Select .ffpfs image…",
+        "v_source_label": "Source Folder (optional — payload comparison)",
+        "v_source_ph": "Optional source folder…",
+        "v_hashes": "Expected Hashes (optional)",
+        "v_crc32": "Expected CRC32 (hex)",
+        "v_crc32_ph": "e.g. 7F528D1F",
+        "v_sha256": "Expected SHA256 (64 hex)",
+        "v_sha256_ph": "64 hex chars…",
+        "v_ekpfs": "EKPFS Key (64 hex, optional)",
+        "v_ekpfs_ph": "64 hex chars…",
+        "v_newcrypt": "Use newCrypt EKPFS derivation",
+        "v_err": "✗ Image path is required.",
+        # Inspect
+        "i_title": "Inspect",
+        "i_subtitle": "View image metadata, version, block size, and warnings",
+        "i_image_label": "PFS Image",
+        "i_image_ph": "Select .ffpfs image…",
+        "i_format": "Output Format",
+        "i_ekpfs": "EKPFS Key (optional)",
+        "i_ekpfs_ph": "64 hex chars…",
+        "i_err": "✗ Image path is required.",
+        # Tree
+        "t_title": "Tree",
+        "t_subtitle": "Browse the directory tree embedded in a PFS image",
+        "t_image_label": "PFS Image",
+        "t_image_ph": "Select .ffpfs image…",
+        "t_ekpfs": "EKPFS Key (optional)",
+        "t_ekpfs_ph": "64 hex chars…",
+        "t_newcrypt": "Use newCrypt EKPFS derivation",
+        "t_err": "✗ Image path is required.",
+        # Unpack
+        "u_title": "Unpack",
+        "u_subtitle": "Extract all files from a PFS image to a destination folder",
+        "u_image_label": "PFS Image",
+        "u_image_ph": "Select .ffpfs image…",
+        "u_out_label": "Output Folder",
+        "u_out_ph": "Destination folder…",
+        "u_overwrite": "Overwrite existing output folder",
+        "u_ekpfs": "EKPFS Key (optional)",
+        "u_ekpfs_ph": "64 hex chars…",
+        "u_newcrypt": "Use newCrypt EKPFS derivation",
+        "u_err": "✗ Image and output folder are both required.",
+        "u_auto_subdir": "→ Output subfolder: {}",
+    },
+    "pt_BR": {
+        "app_subtitle": "Sistema de Arquivos PlayStation",
+        "lang_label": "Idioma",
+        "operations": "OPERAÇÕES",
+        "version_footer": "v{version}",
+        "nav_pack_folder": "📦  Empacotar Pasta",
+        "nav_pack_file": "📄  Empacotar Arquivo",
+        "nav_verify": "✅  Verificar",
+        "nav_inspect": "🔍  Inspecionar",
+        "nav_tree": "🌲  Árvore",
+        "nav_unpack": "📂  Desempacotar",
+        "paths": "Caminhos",
+        "options": "Opções",
+        "encryption": "Criptografia",
+        "output_log": "Log de Saída",
+        "export_log": "Exportar Log",
+        "run": "Executar",
+        "running": "Executando…",
+        "browse": "Procurar",
+        "ok": "✓ Concluído com sucesso.",
+        "err_process": "✗ Processo encerrado com código {}.",
+        "err_not_found": "✗ mkpfs não encontrado. O ambiente virtual está ativo?",
+        "err_unexpected": "[ERRO] Inesperado: {}",
+        "pf_title": "Empacotar Pasta",
+        "pf_subtitle": "Construir uma imagem PFS a partir de um diretório fonte",
+        "pf_src_label": "Pasta Fonte",
+        "pf_src_ph": "Selecione a pasta fonte…",
+        "pf_out_label": "Imagem de Saída",
+        "pf_out_ph": "Selecione o caminho de saída…",
+        "pf_version": "Versão",
+        "pf_compress": "Compressão PFSC",
+        "pf_signed": "Imagem Assinada",
+        "pf_verify": "Verificar Após Empacotar",
+        "pf_dry": "Simulação (Dry Run)",
+        "pf_temp": "Pasta Temporária (opcional)",
+        "pf_temp_ph": "Pasta temp personalizada…",
+        "pkf_temp": "Pasta Temporária (opcional)",
+        "pkf_temp_ph": "Pasta temp personalizada…",
+        "pf_err_src": "✗ A pasta fonte é obrigatória.",
+        "pf_err_out": "✗ O caminho de saída é obrigatório.",
+        "pkf_title": "Empacotar Arquivo",
+        "pkf_subtitle": "Encapsular um único arquivo em uma imagem PFSC comprimida",
+        "pkf_src_label": "Arquivo Fonte",
+        "pkf_src_ph": "Selecione o arquivo fonte…",
+        "pkf_out_label": "Imagem de Saída",
+        "pkf_out_ph": "Selecione o caminho de saída…",
+        "pkf_version": "Versão",
+        "pkf_compress": "Compressão PFSC",
+        "pkf_err": "✗ Arquivo fonte e caminho de saída são obrigatórios.",
+        "v_title": "Verificar",
+        "v_subtitle": "Validar estrutura da imagem, inodes e checksums",
+        "v_image_label": "Imagem PFS",
+        "v_image_ph": "Selecione a imagem .ffpfs…",
+        "v_source_label": "Pasta Fonte (opcional — comparação de payload)",
+        "v_source_ph": "Pasta fonte opcional…",
+        "v_hashes": "Hashes Esperados (opcional)",
+        "v_crc32": "CRC32 Esperado (hex)",
+        "v_crc32_ph": "ex: 7F528D1F",
+        "v_sha256": "SHA256 Esperado (64 hex)",
+        "v_sha256_ph": "64 caracteres hex…",
+        "v_ekpfs": "Chave EKPFS (64 hex, opcional)",
+        "v_ekpfs_ph": "64 caracteres hex…",
+        "v_newcrypt": "Usar derivação newCrypt EKPFS",
+        "v_err": "✗ O caminho da imagem é obrigatório.",
+        "i_title": "Inspecionar",
+        "i_subtitle": "Ver metadados, versão, tamanho de bloco e avisos da imagem",
+        "i_image_label": "Imagem PFS",
+        "i_image_ph": "Selecione a imagem .ffpfs…",
+        "i_format": "Formato de Saída",
+        "i_ekpfs": "Chave EKPFS (opcional)",
+        "i_ekpfs_ph": "64 caracteres hex…",
+        "i_err": "✗ O caminho da imagem é obrigatório.",
+        "t_title": "Árvore",
+        "t_subtitle": "Navegar pela árvore de diretórios embutida na imagem PFS",
+        "t_image_label": "Imagem PFS",
+        "t_image_ph": "Selecione a imagem .ffpfs…",
+        "t_ekpfs": "Chave EKPFS (opcional)",
+        "t_ekpfs_ph": "64 caracteres hex…",
+        "t_newcrypt": "Usar derivação newCrypt EKPFS",
+        "t_err": "✗ O caminho da imagem é obrigatório.",
+        "u_title": "Desempacotar",
+        "u_subtitle": "Extrair todos os arquivos de uma imagem PFS para uma pasta",
+        "u_image_label": "Imagem PFS",
+        "u_image_ph": "Selecione a imagem .ffpfs…",
+        "u_out_label": "Pasta de Destino",
+        "u_out_ph": "Pasta de destino…",
+        "u_overwrite": "Sobrescrever pasta de saída existente",
+        "u_ekpfs": "Chave EKPFS (opcional)",
+        "u_ekpfs_ph": "64 caracteres hex…",
+        "u_newcrypt": "Usar derivação newCrypt EKPFS",
+        "u_err": "✗ Imagem e pasta de destino são obrigatórios.",
+        "u_auto_subdir": "→ Subpasta de saída: {}",
+    },
+    "es": {
+        "app_subtitle": "Sistema de Archivos PlayStation",
+        "lang_label": "Idioma",
+        "operations": "OPERACIONES",
+        "version_footer": "v{version}",
+        "nav_pack_folder": "📦  Empaquetar Carpeta",
+        "nav_pack_file": "📄  Empaquetar Archivo",
+        "nav_verify": "✅  Verificar",
+        "nav_inspect": "🔍  Inspeccionar",
+        "nav_tree": "🌲  Árbol",
+        "nav_unpack": "📂  Desempaquetar",
+        "paths": "Rutas",
+        "options": "Opciones",
+        "encryption": "Cifrado",
+        "output_log": "Registro de Salida",
+        "export_log": "Exportar Registro",
+        "run": "Ejecutar",
+        "running": "Ejecutando…",
+        "browse": "Buscar",
+        "ok": "✓ Completado con éxito.",
+        "err_process": "✗ El proceso terminó con código {}.",
+        "err_not_found": "✗ mkpfs no encontrado. ¿El entorno virtual está activo?",
+        "err_unexpected": "[ERROR] Inesperado: {}",
+        "pf_title": "Empaquetar Carpeta",
+        "pf_subtitle": "Construir una imagen PFS desde un directorio fuente",
+        "pf_src_label": "Carpeta Fuente",
+        "pf_src_ph": "Seleccionar carpeta fuente…",
+        "pf_out_label": "Imagen de Salida",
+        "pf_out_ph": "Seleccionar ruta de salida…",
+        "pf_version": "Versión",
+        "pf_compress": "Compresión PFSC",
+        "pf_signed": "Imagen Firmada",
+        "pf_verify": "Verificar Tras Empaquetar",
+        "pf_dry": "Simulación (Dry Run)",
+        "pf_temp": "Carpeta Temporal (opcional)",
+        "pf_temp_ph": "Carpeta temp personalizada…",
+        "pkf_temp": "Carpeta Temporal (opcional)",
+        "pkf_temp_ph": "Carpeta temp personalizada…",
+        "pf_err_src": "✗ La carpeta fuente es obligatoria.",
+        "pf_err_out": "✗ La ruta de salida es obligatoria.",
+        "pkf_title": "Empaquetar Archivo",
+        "pkf_subtitle": "Envolver un solo archivo en una imagen PFSC comprimida",
+        "pkf_src_label": "Archivo Fuente",
+        "pkf_src_ph": "Seleccionar archivo fuente…",
+        "pkf_out_label": "Imagen de Salida",
+        "pkf_out_ph": "Seleccionar ruta de salida…",
+        "pkf_version": "Versión",
+        "pkf_compress": "Compresión PFSC",
+        "pkf_err": "✗ El archivo fuente y la ruta de salida son obligatorios.",
+        "v_title": "Verificar",
+        "v_subtitle": "Validar estructura de imagen, inodes y checksums de payload",
+        "v_image_label": "Imagen PFS",
+        "v_image_ph": "Seleccionar imagen .ffpfs…",
+        "v_source_label": "Carpeta Fuente (opcional — comparación de payload)",
+        "v_source_ph": "Carpeta fuente opcional…",
+        "v_hashes": "Hashes Esperados (opcional)",
+        "v_crc32": "CRC32 Esperado (hex)",
+        "v_crc32_ph": "ej: 7F528D1F",
+        "v_sha256": "SHA256 Esperado (64 hex)",
+        "v_sha256_ph": "64 caracteres hex…",
+        "v_ekpfs": "Clave EKPFS (64 hex, opcional)",
+        "v_ekpfs_ph": "64 caracteres hex…",
+        "v_newcrypt": "Usar derivación newCrypt EKPFS",
+        "v_err": "✗ La ruta de imagen es obligatoria.",
+        "i_title": "Inspeccionar",
+        "i_subtitle": "Ver metadatos, versión, tamaño de bloque y advertencias",
+        "i_image_label": "Imagen PFS",
+        "i_image_ph": "Seleccionar imagen .ffpfs…",
+        "i_format": "Formato de Salida",
+        "i_ekpfs": "Clave EKPFS (opcional)",
+        "i_ekpfs_ph": "64 caracteres hex…",
+        "i_err": "✗ La ruta de imagen es obligatoria.",
+        "t_title": "Árbol",
+        "t_subtitle": "Explorar el árbol de directorios embebido en una imagen PFS",
+        "t_image_label": "Imagen PFS",
+        "t_image_ph": "Seleccionar imagen .ffpfs…",
+        "t_ekpfs": "Clave EKPFS (opcional)",
+        "t_ekpfs_ph": "64 caracteres hex…",
+        "t_newcrypt": "Usar derivación newCrypt EKPFS",
+        "t_err": "✗ La ruta de imagen es obligatoria.",
+        "u_title": "Desempaquetar",
+        "u_subtitle": "Extraer todos los archivos de una imagen PFS a una carpeta",
+        "u_image_label": "Imagen PFS",
+        "u_image_ph": "Seleccionar imagen .ffpfs…",
+        "u_out_label": "Carpeta de Destino",
+        "u_out_ph": "Carpeta de destino…",
+        "u_overwrite": "Sobrescribir carpeta de salida existente",
+        "u_ekpfs": "Clave EKPFS (opcional)",
+        "u_ekpfs_ph": "64 caracteres hex…",
+        "u_newcrypt": "Usar derivación newCrypt EKPFS",
+        "u_err": "✗ La imagen y la carpeta de destino son obligatorias.",
+        "u_auto_subdir": "→ Subcarpeta de salida: {}",
+    },
+}
+
+_LANG_NAMES: dict[str, str] = {
+    "en": "English",
+    "pt_BR": "Português (BR)",
+    "es": "Español",
+}
+
+# Active locale — updated by the language selector
+_current_locale: str = "en"
+
+
+def set_locale(locale: str) -> None:
+    """Set the active GUI locale.
+
+    Args:
+        locale: Locale key to activate (must exist in _TRANSLATIONS).
+    """
+    global _current_locale
+    _current_locale = locale
+
+
+def tr(key: str) -> str:
+    """Return the translated string for the given key in the active locale.
+
+    Args:
+        key: Translation dictionary key.
+
+    Returns:
+        Translated string, falling back to English when the key is missing.
+    """
+    from mkpfs import __version__
+
+    lang: dict[str, str] = _TRANSLATIONS.get(_current_locale, _TRANSLATIONS["en"])
+    value: str = lang.get(key, _TRANSLATIONS["en"].get(key, key))
+    return value.replace("{version}", __version__)
