+++ mkpfs/gui/panels/__init__.py
@@ -1,0 +1,11 @@
+"""Panel package for the mkpfs GUI - hosts the base panel and operation panels."""
+
+from .base import BasePanel
+from .inspect import InspectPanel
+from .pack_file import PackFilePanel
+from .pack_folder import PackFolderPanel
+from .tree import TreePanel
+from .unpack import UnpackPanel
+from .verify import VerifyPanel
+
+__all__ = ["BasePanel", "InspectPanel", "PackFilePanel", "PackFolderPanel", "TreePanel", "UnpackPanel", "VerifyPanel"]
